+++ ReleaseNotes/release_notes.md
@@ -21,11 +21,12 @@
   1. add `Bernstein3D::integralXY`,`Bernstein3D::integralXZ` and `Bernstein3D::integralYZ` methods
   1. add `Bernstein3D::integralX`,`Bernstein3D::integralY` and `Bernstein3D::integralZ` methods
   1. extend tests for `Bernstein3D` objects
-  
+  1. allow to define the default tabel style either via connfigiraitno file (section 'Tables', field `Style`) or envitonment variable `OSTAP_TABLE_STYLE`
   
 ## Backward incompatible:  
 
   1. rename `BernsteinDualBasis` into `Ostap::Math::BernsteinnDual`
+  1. `ostap.logger.table.table` : rename argument `format` to `style`
   
 ## Bug fixes:
 
@@ -50,7 +51,7 @@
 ## Backward incompatible:  
 
   1. rename some methods for `Ostap::Math::Integrator`
-
+  
 ## Bug fixes:
 
   1. Fix typos for `Ostap::Math::Integrator`